+++ client/src/components/home/Hero/Hero.tsx
@@ -1,5 +1,5 @@
 import { Button } from '@/components/ui';
-import { AUTH_PAGES } from '@/config/authPages.config';
+import { CROSS_DOMAIN_ROUTES } from '@/config/navigation.config';
 import styles from './Hero.module.scss';
 
 export const Hero = () => {
@@ -16,7 +16,9 @@ export const Hero = () => {
 						Global distributed checks, hybrid project grouping, and public
 						status pages. Forever free.
 					</p>
-					<Button href={AUTH_PAGES.REGISTER}>Start Monitoring Now</Button>
+					<Button href={CROSS_DOMAIN_ROUTES.APP_REGISTER} asNative>
+						Start Monitoring Now
+					</Button>
 				</div>
 			</div>
 		</div>