+++ client/src/components/layout/Header/HeaderMain/HeaderMain.tsx
@@ -4,7 +4,6 @@ import { Logo } from '@/components/icons/Logo';
 import { Button } from '@/components/ui';
 import { CROSS_DOMAIN_ROUTES } from '@/config/navigation.config';
 import { PUBLIC_PAGES } from '@/config/publicPages.config';
-import Link from 'next/link';
 import { HeaderContainer } from '../HeaderContainer';
 import styles from './HeaderMain.module.scss';
 
@@ -14,18 +13,25 @@ export const HeaderMain = () => {
 			headerClassName={styles.header}
 			containerClassName={styles.container}
 		>
-			<Link href={PUBLIC_PAGES.HOME} className={styles.logo}>
+			<a href={PUBLIC_PAGES.HOME} className={styles.logo}>
 				<Logo width={32} height={32} />
 				<span className={styles.logoText}>
 					<span className={styles.highlight}>Site</span>
 					Status
 				</span>
-			</Link>
+			</a>
 			<div className={styles.buttons}>
-				<Button variant="link" href={CROSS_DOMAIN_ROUTES.APP_LOGIN} size="sm">
+				<Button
+					variant="link"
+					href={CROSS_DOMAIN_ROUTES.APP_LOGIN}
+					size="sm"
+					asNative
+				>
 					Log in
 				</Button>
-				<Button href={CROSS_DOMAIN_ROUTES.APP_REGISTER}>Sign Up</Button>
+				<Button href={CROSS_DOMAIN_ROUTES.APP_REGISTER} asNative>
+					Sign Up
+				</Button>
 			</div>
 		</HeaderContainer>
 	);