+++ client/src/components/monitors/Monitors/MonitorsAll.tsx
@@ -2,6 +2,7 @@
 
 import { getAllMonitors } from '@/api/monitor/getAllMonitors.api';
 import { IconButton, SectionHeader } from '@/components/ui';
+import { PRIVATE_PAGES } from '@/config/privatePages.config';
 import { QUERY_KEYS } from '@/config/queryClient.config';
 import { FiPlus } from 'react-icons/fi';
 import { Monitors } from './Monitors';
@@ -12,7 +13,11 @@ export const MonitorsAll = () => {
 			<SectionHeader
 				title="Monitors"
 				description="View and manage your monitors"
-				rightSlot={<IconButton Icon={FiPlus}>Add New Monitor</IconButton>}
+				rightSlot={
+					<IconButton Icon={FiPlus} href={PRIVATE_PAGES.MONITORS.NEW}>
+						Add New Monitor
+					</IconButton>
+				}
 			/>
 			<Monitors
 				queryKeyBase={QUERY_KEYS.monitors.lists()}