+++ server/src/app.module.ts
@@ -10,15 +10,9 @@ import { AppController } from './app.controller';
 @Module({
 	imports: [
 		ConfigModule.forRoot({ isGlobal: true }),
-		// ThrottlerModule.forRoot([
-			// { name: 'short', ttl: 1000, limit: 10 },
-			// { name: 'default', ttl: 60000, limit: 100 },
-			// { name: 'strict', ttl: 60000, limit: 5 },
-		// ]),
 		ThrottlerModule.forRoot([
-			{ name: 'short', ttl: 1000, limit: 1000000 },
-			{ name: 'default', ttl: 60000, limit: 1000000 },
-			{ name: 'strict', ttl: 60000, limit: 500000 },
+			{ name: 'short', ttl: 1000, limit: 20 },
+			{ name: 'default', ttl: 60000, limit: 200 },
 		]),
 		EventEmitterModule.forRoot(),
 		InfraModule,