+++ server/src/config/swagger/public-swagger.config.ts
@@ -4,5 +4,5 @@ export const getPublicSwaggerConfig = () =>
 	basicSwaggerConfig(
 		'SiteStatus API',
 		'SiteStatus API documentation.',
-		'1.0.0',
+		'1.2.0',
 	).build();